change et-common dependency
