--- src/main/java/uk/gov/hmcts/reform/et/syaapi/model/helper/Constants.java
@@ -1,12 +1,0 @@
-package uk.gov.hmcts.reform.et.syaapi.model.helper;
-
-public class Constants {
-
-    private Constants() {
-        // Class only provides constant values
-    }
-
-    public static final String ENGLANDWALES_CASE_TYPE_ID = "ET_EnglandWales";
-    public static final String SCOTLAND_CASE_TYPE_ID = "ET_Scotland";
-
-}
