--- src/main/java/uk/gov/hmcts/reform/et/syaapi/service/exceptions/PostcodeNotFoundInLookupException.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.et.syaapi.service.exceptions;
-
-public class PostcodeNotFoundInLookupException extends Exception {
-    public PostcodeNotFoundInLookupException(String input) {
-        super(input);
-    }
-}
